Convert addon to work with Statamic v3, and update readme installation instructions
