--- svg/composer.json
@@ -1,8 +1,0 @@
-{
-  "name": "statamic-svg",
-  "autoload": {
-    "classmap": [
-      "./"
-    ]
-  }
-}
