--- svg/meta.yaml
@@ -1,6 +1,0 @@
-name: SVG
-version: "2.0"
-description: 
-url: https://github.com/benfurfie/statamic-svg.git
-developer: Ben Furfie
-developer_url: https://www.benfurfie.co.uk
