--- svg/settings.yaml
@@ -1,4 +1,0 @@
-fields:
-  default_asset_container:
-    default: main
-    type: asset_container
